:art: Move to init
