+++ lamindb/__init__.py
@@ -96,7 +96,7 @@ if _check_instance_setup(from_module="lnschema_core"):
     from ._view import view
     from .core._context import context
     from .core._settings import settings
-    from .curators._curators import Curator
+    from .curators import Curator
 
     track = context.track  # simple access because these are so common
     finish = context.finish  # simple access because these are so common
