+++ lamindb/core/__init__.py
@@ -85,7 +85,7 @@ from lamindb._query_manager import QueryManager
 from lamindb._query_set import QuerySet, RecordList
 from lamindb.core._feature_manager import FeatureManager, ParamManager
 from lamindb.core._label_manager import LabelManager
-from lamindb.curators._curators import (
+from lamindb.curators import (
     AnnDataCurator,
     BaseCurator,
     CurateLookup,
