+++ lamindb/curators/_spatial.py
@@ -5,11 +5,15 @@ from typing import Any
 import lamindb_setup as ln_setup
 import pandas as pd
 from lamin_utils import colors, logger
+from lnschema_core.models import Artifact, Collection, Feature, FeatureSet, Record, Run
+from lnschema_core.types import FieldAttr
+from spatialdata import SpatialData
+
 from lamindb.core._data import add_labels
 from lamindb.core._feature_manager import parse_feature_sets_from_anndata
 from lamindb.core._settings import settings
 from lamindb.core.exceptions import ValidationError
-from lamindb.curators._curators import (
+from lamindb.curators import (
     AnnDataCurator,
     CurateLookup,
     DataFrameCurator,
@@ -18,9 +22,6 @@ from lamindb.curators._curators import (
     check_registry_organism,
     get_current_filter_kwargs,
 )
-from lnschema_core.models import Artifact, Collection, Feature, FeatureSet, Record, Run
-from lnschema_core.types import FieldAttr
-from spatialdata import SpatialData
 
 
 class SpatialDataCurator:
